Fix notifications.
