+++ src/android/main.c
@@ -4,6 +4,7 @@
 #include "../settings.h"
 
 #include "../native/keyboard.h"
+#include "../native/notify.h"
 
 #include <pthread.h>
 #include <stdint.h>
@@ -28,7 +29,8 @@ static void freefonts(void);
 #include "gl.c"
 
 
-static volatile bool destroy, focused;
+static volatile bool destroy;
+bool have_focus = false;
 
 static bool shift;
 
@@ -805,7 +807,7 @@ static void android_main(struct android_app *state) {
             }
         }
 
-        if (window != NULL && focused) {
+        if (window != NULL && have_focus) {
             utox_android_redraw_window();
         }
 
@@ -876,7 +878,7 @@ static void onNativeWindowDestroyed(ANativeActivity *act, ANativeWindow *win) {
 }
 
 static void onWindowFocusChanged(ANativeActivity *act, int focus) {
-    focused = (focus != 0);
+    have_focus = (focus != 0);
 }
 
 static void onInputQueueCreated(ANativeActivity *act, AInputQueue *queue) {
