+++ src/windows/notify.c
@@ -1,14 +1,52 @@
 #include "notify.h"
 
 #include "main.h"
+#include "utf8.h"
 #include "window.h"
 
 #include "../debug.h"
+#include "../macros.h"
+#include "../self.h"
 #include "../ui.h"
 
+#include "../native/notify.h"
 #include "../native/window.h"
 
 #include <windowsx.h>
+
+bool have_focus = false;
+
+/** Creates a tray baloon popup with the message, and flashes the main window
+ *
+ * accepts: char *title, title length, char *msg, msg length;
+ * returns void;
+ */
+void notify(char *title, uint16_t title_length, const char *msg, uint16_t msg_length, void *UNUSED(object), bool UNUSED(is_group)) {
+    if (have_focus || self.status == 2) {
+        return;
+    }
+
+    FlashWindow(main_window.window, true);
+    flashing = true;
+
+    NOTIFYICONDATAW nid = {
+        .cbSize      = sizeof(nid),
+        .hWnd        = main_window.window,
+        .uFlags      = NIF_ICON | NIF_INFO,
+        .hIcon       = unread_messages_icon,
+        .uTimeout    = 5000,
+        .dwInfoFlags = 0,
+    };
+
+    utf8tonative(title, nid.szInfoTitle, title_length > sizeof(nid.szInfoTitle) / sizeof(*nid.szInfoTitle) - 1 ?
+                                             sizeof(nid.szInfoTitle) / sizeof(*nid.szInfoTitle) - 1 :
+                                             title_length);
+    utf8tonative(msg, nid.szInfo, msg_length > sizeof(nid.szInfo) / sizeof(*nid.szInfo) - 1 ?
+                                      sizeof(nid.szInfo) / sizeof(*nid.szInfo) - 1 :
+                                      msg_length);
+
+    Shell_NotifyIconW(NIM_MODIFY, &nid);
+}
 
 static void redraw_notify(UTOX_WINDOW *win) {
     LOG_TRACE("Notify", "redraw start");
