+++ src/xlib/event.c
@@ -19,6 +19,7 @@
 
 #include "../native/clipboard.h"
 #include "../native/keyboard.h"
+#include "../native/notify.h"
 #include "../native/ui.h"
 
 #include "../ui/draw.h" // Needed for enddraw. This should probably be changed.
@@ -37,6 +38,8 @@
 #include "../main.h" // STBI
 
 extern XIC xic;
+
+bool have_focus = false;
 
 static void mouse_move(XMotionEvent *event, UTOX_WINDOW *window) {
     if (pointergrab) { // TODO super globals are bad mm'kay?
@@ -309,7 +312,7 @@ bool doevent(XEvent event) {
             }
             #endif
 
-            havefocus      = true;
+            have_focus = true;
             XWMHints hints = { 0, 0, 0, 0, 0, 0, 0, 0, 0 };
             XSetWMHints(display, main_window.window, &hints);
             break;
@@ -326,7 +329,7 @@ bool doevent(XEvent event) {
             }
             #endif
 
-            havefocus = false;
+            have_focus = false;
             break;
         }
