+++ src/xlib/main.c
@@ -22,6 +22,7 @@
 #include "../av/utox_av.h"
 
 #include "../native/image.h"
+#include "../native/notify.h"
 #include "../native/ui.h"
 
 #include "../ui/draw.h"
@@ -70,9 +71,10 @@ void postmessage_utox(UTOX_MSG msg, uint16_t param1, uint16_t param2, void *data
     XFlush(display);
 }
 
-FILE    *ptt_keyboard_handle;
-Display *ptt_display;
-void     init_ptt(void) {
+static FILE *   ptt_keyboard_handle;
+static Display *ptt_display;
+
+void init_ptt(void) {
     settings.push_to_talk = 1;
 
     char path[UTOX_FILE_NAME_LENGTH];
@@ -575,7 +577,7 @@ void setscale_fonts(void) {
 }
 
 void notify(char *title, uint16_t UNUSED(title_length), const char *msg, uint16_t msg_length, void *object, bool is_group) {
-    if (havefocus) {
+    if (have_focus) {
         return;
     }
